Also check for effective uid (paranoia).
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@140 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ appl/xnlock/xnlock.c
@@ -542,7 +542,8 @@ GetPasswd(Widget w, XEvent *_event, String *_s, Cardinal *_n)
 	/*
 	 * Password that log out user
 	 */
-	if (getuid() != 0
+	if (    getuid() != 0
+	    && geteuid() != 0
 	    && (time(0) - locked_at) > ALLOW_LOGOUT
 	    && strncmp(passwd, LOGOUT_PASSWD, sizeof(LOGOUT_PASSWD)) == 0)
 	    kill(-1, SIGHUP);
